requirements.txt
bidict==0.21.2
gunicorn==20.0.4
python-engineio==4.0.0
python-socketio==5.0.4
eventlet==0.24.1
flask-socketio==5.1.2
flask==2.1.2
mysql-connector-python==8.0.24 
